Fully functional marketplace with filters and sorting capabilities
