+++ src/components/MxCatalog/Explore/SoldOnceNFT.js
@@ -4,17 +4,17 @@ const SoldOnceNFT = ({ track }) => {
 	if (track) {
 		const soldOnceTrackData = {
 			// This is passed only because we need Lowest Price being shown on the NFT Card
-			tokenIdHavingLowestPrice: track.purchasedTokens.at(0),
+			tokenIdHavingLowestPrice: track.tokenIdHavingLowestPrice,
 		};
 
 		return (
 			<NFTCard
-				redirectLink={`/track/polygon/${track.purchasedTokens.at(0)}`}
+				redirectLink={`/track/polygon/${track.tokenIdHavingLowestPrice}`}
 				trackName={track.title}
 				artistName={track.artist}
 				artistAddress={track.artistAddress}
 				image={track.artwork.uri.replace("ipfs://", process.env.NEXT_PUBLIC_IPFS_NODE_URL)}
-				tokenId={track.purchasedTokens.at(0)}
+				tokenId={track.tokenIdHavingLowestPrice}
 				numberOfCopies={track.numberOfCopies}
 				collaboratorList={track.collaborators}
 				soldOnceTrackData={soldOnceTrackData}
