+++ src/components/MxCatalog/Utils/CatalogBody.js
@@ -21,7 +21,7 @@ export default function CatalogBody({ currentSelection, setAppliedFilter, applie
 					</div>
 
 					<div className="flex items-start justify-between space-x-10 w-full">
-						<Filter setAppliedFilter={setAppliedFilter} appliedFilter={appliedFilter} />
+						<Filter setAppliedFilter={setAppliedFilter} appliedFilter={appliedFilter} currentSelection={currentSelection} />
 
 						<div className="w-full grid col-span-9 gap-y-[60px] gap-x-[80px] 2xl:grid-cols-4 xl:grid-cols-3 md:grid-cols-2">
 							{currentSelection === 1 ? (
